Checkout These Talons! (#1775)
* Remove unnecessary cart component from checkout. Remove unnecessary props from receipt args

* Export functionality to talons and use within checkout components

* trim the talons

* Oops forgot some edits

* Move anchorRef back to component and install informed peer dependency

* Wrap callbacks in talons before returning to component

* Update yarn lock because of informed peer dep I guess
