--- packages/venia-ui/lib/components/Checkout/cart.css
@@ -1,3 +1,0 @@
-.root {
-    composes: footer from './flow.css';
-}
