+++ packages/venia-ui/lib/components/Checkout/flow.js
@@ -2,62 +2,53 @@ import React, { useCallback } from 'react';
 import { func, shape, string } from 'prop-types';
 
 import { mergeClasses } from '../../classify';
-import Cart from './cart';
+import CheckoutButton from './checkoutButton';
 import Form from './form';
 import Receipt from './Receipt';
 import defaultClasses from './flow.css';
 import isObjectEmpty from '../../util/isObjectEmpty';
-import { useCartContext } from '@magento/peregrine/lib/context/cart';
-import { useCheckoutContext } from '@magento/peregrine/lib/context/checkout';
-import { useUserContext } from '@magento/peregrine/lib/context/user';
 import { useToasts } from '@magento/peregrine';
 import Icon from '../Icon';
 
 import { AlertCircle as AlertCircleIcon } from 'react-feather';
+import { useFlow } from '@magento/peregrine/lib/talons/Checkout/useFlow';
 const ErrorIcon = <Icon src={AlertCircleIcon} attrs={{ width: 18 }} />;
-
-const isCartReady = cart => cart.details && cart.details.items_count > 0;
-const isCheckoutReady = checkout => {
-    const {
-        billingAddress,
-        paymentData,
-        shippingAddress,
-        shippingMethod
-    } = checkout;
-
-    const objectsHaveData = [
-        billingAddress,
-        paymentData,
-        shippingAddress
-    ].every(data => {
-        return !!data && !isObjectEmpty(data);
-    });
-
-    const stringsHaveData = !!shippingMethod && shippingMethod.length > 0;
-
-    return objectsHaveData && stringsHaveData;
-};
 
 /**
  * This Flow component's primary purpose is to take relevant state and actions
  * and pass them to the current checkout step.
  */
 const Flow = props => {
-    const [cart] = useCartContext();
-    const [
-        checkoutState,
-        {
-            beginCheckout,
-            cancelCheckout,
-            submitOrder,
-            submitPaymentMethodAndBillingAddress,
-            submitShippingAddress,
-            submitShippingMethod
-        }
-    ] = useCheckoutContext();
-    const [user] = useUserContext();
+    const { step } = props;
+    const [, { addToast }] = useToasts();
+    const onSubmitError = useCallback(() => {
+        addToast({
+            type: 'error',
+            icon: ErrorIcon,
+            message:
+                'Something went wrong submitting your order! Try again later.',
+            timeout: 7000
+        });
+    }, [addToast]);
 
-    const { step, setStep } = props;
+    const talonProps = useFlow({
+        onSubmitError,
+        setStep: props.setStep
+    });
+
+    const {
+        cartState,
+        checkoutDisabled,
+        checkoutState,
+        isReady,
+        submitPaymentMethodAndBillingAddress,
+        submitShippingAddress,
+        submitShippingMethod,
+        handleBeginCheckout,
+        handleCancelCheckout,
+        handleCloseReceipt,
+        handleSubmitOrder
+    } = talonProps;
 
     const {
         availableShippingMethods,
@@ -73,44 +64,16 @@ const Flow = props => {
     const classes = mergeClasses(defaultClasses, props.classes);
 
     let child;
-    const [, { addToast }] = useToasts();
-    const handleBeginCheckout = useCallback(async () => {
-        await beginCheckout();
-        setStep('form');
-    }, [beginCheckout, setStep]);
-
-    const handleCancelCheckout = useCallback(async () => {
-        await cancelCheckout();
-        setStep('cart');
-    }, [cancelCheckout, setStep]);
-
-    const handleSubmitOrder = useCallback(async () => {
-        try {
-            await submitOrder();
-            setStep('receipt');
-        } catch (e) {
-            addToast({
-                type: 'error',
-                icon: ErrorIcon,
-                message:
-                    'Something went wrong submitting your order! Try again later.',
-                timeout: 7000
-            });
-        }
-    }, [addToast, setStep, submitOrder]);
-
-    const handleCloseReceipt = useCallback(() => {
-        setStep('cart');
-    }, [setStep]);
-
     switch (step) {
         case 'cart': {
-            const stepProps = {
-                beginCheckout: handleBeginCheckout,
-                ready: !isSubmitting && isCartReady(cart)
-            };
-
-            child = <Cart {...stepProps} />;
+            child = (
+                <div className={classes.footer}>
+                    <CheckoutButton
+                        disabled={checkoutDisabled}
+                        onClick={handleBeginCheckout}
+                    />
+                </div>
+            );
             break;
         }
         case 'form': {
@@ -118,7 +81,7 @@ const Flow = props => {
                 availableShippingMethods,
                 billingAddress,
                 cancelCheckout: handleCancelCheckout,
-                cart,
+                cart: cartState,
                 checkout: checkoutState,
                 hasPaymentMethod: !!paymentData && !isObjectEmpty(paymentData),
                 hasShippingAddress:
@@ -127,7 +90,7 @@ const Flow = props => {
                     !!shippingMethod && !isObjectEmpty(shippingMethod),
                 isSubmitting,
                 paymentData,
-                ready: isCheckoutReady(checkoutState),
+                ready: isReady,
                 shippingAddress,
                 shippingAddressError,
                 shippingMethod,
@@ -142,12 +105,7 @@ const Flow = props => {
             break;
         }
         case 'receipt': {
-            const stepProps = {
-                user,
-                onClose: handleCloseReceipt
-            };
-
-            child = <Receipt {...stepProps} />;
+            child = <Receipt onClose={handleCloseReceipt} />;
             break;
         }
         default: {
